+++ sphinx_needs/directives/needuml.py
@@ -75,10 +75,7 @@ class NeedumlDirective(Directive):
             if os.path.isabs(save_path):
                 raise NeedumlException(f"Given save path: {save_path}, is not relative path.")
             else:
-                plantuml_code_out_path = os.path.join(env.app.confdir, save_path)
-                save_dir = os.path.dirname(plantuml_code_out_path)
-                if not os.path.exists(save_dir):
-                    os.makedirs(save_dir, exist_ok=True)
+                plantuml_code_out_path = save_path
 
         env.needs_all_needumls[targetid] = {
             "docname": env.docname,
@@ -219,15 +216,13 @@ def process_needuml(app, doctree, fromdocname):
 
         puml_node["uml"] += "\n@enduml\n"
 
+        # Add calculated needuml content
+        current_needuml["content_calculated"] = puml_node["uml"]
+
         puml_node["incdir"] = os.path.dirname(current_needuml["docname"])
         puml_node["filename"] = os.path.split(current_needuml["docname"])[1]  # Needed for plantuml >= 0.9
 
         content.append(puml_node)
-
-        # save the generated plantuml-code into given file path, e.g. my_needuml.puml
-        if current_needuml["save"]:
-            with open(current_needuml["save"], "w") as f:
-                f.write(puml_node["uml"])
 
         if current_needuml["caption"]:
             title_text = current_needuml["caption"]