+++ sphinx_needs/needs.py
@@ -9,7 +9,12 @@ from sphinx.errors import SphinxError
 from sphinx.roles import XRefRole
 
 from sphinx_needs.api.configuration import add_extra_option
-from sphinx_needs.builder import NeedsBuilder, build_needs_json
+from sphinx_needs.builder import (
+    NeedsBuilder,
+    NeedumlsBuilder,
+    build_needs_json,
+    build_needumls_pumls,
+)
 from sphinx_needs.config import NEEDS_CONFIG
 from sphinx_needs.defaults import (
     DEFAULT_DIAGRAM_TEMPLATE,
@@ -108,6 +113,7 @@ def setup(app: Sphinx) -> Dict[str, Any]:
     app.setup_extension("sphinx_data_viewer")
 
     app.add_builder(NeedsBuilder)
+    app.add_builder(NeedumlsBuilder)
     app.add_config_value(
         "needs_types",
         [
@@ -214,6 +220,8 @@ def setup(app: Sphinx) -> Dict[str, Any]:
     app.add_config_value("needs_string_links", {}, "html", types=[dict])
 
     app.add_config_value("needs_build_json", False, "html", types=[bool])
+
+    app.add_config_value("needs_build_needumls", "", "html", types=[str])
 
     # Permalink related config values.
     # path to permalink.html; absolute path from web-root
@@ -325,6 +333,7 @@ def setup(app: Sphinx) -> Dict[str, Any]:
     app.connect("doctree-resolved", process_need_func)
     app.connect("build-finished", process_warnings)
     app.connect("build-finished", build_needs_json)
+    app.connect("build-finished", build_needumls_pumls)
     app.connect("env-updated", install_lib_static_files)
     app.connect("env-updated", install_permalink_file)